pyproject.toml
[tool.poetry]
name = "ai-voice-assistant"
version = "0.1.0"
description = ""
authors = ["Mic <mdaizovi@gmail.com>"]
readme = "README.md"
packages = [{include = "ai_voice_assistant"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.99.1"
uvicorn = "^0.22.0"
openai = "^0.27.8"
twilio = "^8.4.0"
python-decouple = "^3.8"
python-multipart = "^0.0.6"
pyngrok = "^6.0.0"
pydub = "^0.25.1"
requests = "^2.31.0"
boto3 = "^1.26.165"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
